delete button now causes render of interested event list
